refactor(EntryController->delete): rewrite it to be reuseable

+++ tools/bazar/controllers/EntryController.php
@@ -4,6 +4,8 @@ namespace YesWiki\Bazar\Controller;
 
 use DateInterval;
 use DateTime;
+use Exception;
+use Throwable;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use YesWiki\Bazar\Exception\UserFieldException;
 use YesWiki\Bazar\Field\BazarField;
@@ -349,13 +351,31 @@ class EntryController extends YesWikiController
         ]);
     }
 
-    public function delete($entryId)
+    public function delete($entryId, bool $redirectAfter = false): bool
     {
-        $this->triggerDeletedEventIfNeeded(function()use($entryId){
-            $this->entryManager->delete($entryId);
-        },$entryId);
-        flash(_t('BAZ_FICHE_SUPPRIMEE')." ($entryId)" , 'success');
-        $this->wiki->Redirect($this->wiki->Href('', 'BazaR', ['vue' => 'consulter'],false));
+        if ($this->entryManager->isEntry($entryId)){
+            try {
+                $entry = $this->entryManager->getOne($entryId);
+                $this->entryManager->delete($entryId);
+                if(!$this->entryManager->isEntry($entryId)){
+                    $this->triggerDeletedEvent($entryId,$entry);
+                    if ($redirectAfter){
+                        flash(_t('BAZ_FICHE_SUPPRIMEE')." ($entryId)" , 'success');
+                        $this->wiki->Redirect($this->wiki->Href('', 'BazaR', ['vue' => 'consulter'],false));
+                    }
+                    return true;
+                }
+            } catch (Throwable $th) {
+                if ($redirectAfter){
+                    flash(_t('DELETEPAGE_NOT_DELETED')." ($entryId) : {$th->getMessage()}" , 'error');
+                    $this->wiki->Redirect($this->wiki->Href('', 'BazaR', ['vue' => 'consulter'],false));
+                }
+                throw new Exception($th->getMessage(),$th->getCode(),$th);
+            }
+            return false;
+        } else {
+            throw new Exception('Not deleted because not entry'.(is_scalar($entryId) ? ' ('.strval($entryId).')' : ''));
+        }
     }
 
     public function triggerDeletedEventIfNeeded($callback,$entryId)
